change to long index for bucketmapper

+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/aggregator/BucketMapper.java
@@ -11,10 +11,11 @@ package io.opentelemetry.sdk.metrics.internal.aggregator;
  */
 interface BucketMapper {
   /**
-   * Maps a recorded double value to a bucket index.
+   * Maps a recorded double value to a bucket index. If the index falls out of the range that can
+   * be represented by an int, then it is expected that the histogram would be downscaled.
    *
    * @param value Measured value
    * @return the index of the bucket which the value maps to.
    */
-  int valueToIndex(double value);
+  long valueToIndex(double value);
 }
